HKXWriter bug fixed and code tidy-up
Fixed a positioning bug in HKXWriter usage of the ByteBuffer before
writing to file that caused an invalid file format.

1 Added and amended javadocs to all methods altered to use ByteBuffers
instead of File to make the usage clearer
2 Deprecated some close() and a clean() method to indicate to third
parties they are not useful.
3 Removed many throws IOException and a few throws FileNotFoundException
as they were no longer in fact thrown (ByteBuffer tends to throw
IndexOutOfBoundException) and removed the javadoc of that exception
throwing
4 Removed unused import statements

Note:
Did not update variable names, but marked the type as ByteBuffer in the
javadoc
Did not add javadoc to classes that didn't have it already, so as not to
mis-lead regarding the usage and function of the class