Some more PMD goodness. Fixed a bunch of little warnings, a lot more to go.
The ruleset is probably the one that will be kept for this project, as most of the rules are relevant. Feel free to "un-exceptionize" the rules you want, but resolving these excluded rules aren't valid reasons for a change to the codebase (you need another reason). Except Demeter, it isn't friend with the Dom elements.